Reorganize Game components
Extract top-level `Game` component from `game/page.tsx`.

Move files from `components/game` to `game/{ui,3d}`.

Remove superfluous `"use client";` directives except for the top level
entry point component `Game`. It is only needed at the boundary between
server components and client components, and creates incorrect warnings
from Next.js TypeScript plugin when passing non-serializable data
between client components.

Combine `GameCanvas` and `PongScheme` into `GameRender`. The separation
into two components is an implementation detail, so keep the inner
component private. Move state and scene to the outer component to keep
the inner component minimal.
