+++ src/game/3d/Arena.tsx
