+++ src/game/3d/Paddle.tsx
