+++ src/game/ui/ConfirmModal.tsx
